+++ libs/ui/src/lib/line-chart/line-chart.component.ts
@@ -25,6 +25,7 @@ import {
 } from '@angular/core';
 import {
   Chart,
+  Decimation,
   Filler,
   LineController,
   LineElement,
@@ -46,6 +47,7 @@ export class LineChartComponent implements AfterViewInit, OnChanges, OnDestroy {
   @Input() benchmarkLabel = '';
   @Input() colorScheme: ColorScheme;
   @Input() currency: string;
+  @Input() dataDecimation = false;
   @Input() historicalDataItems: LineChartItem[];
   @Input() isAnimated = false;
   @Input() locale = getLocale();
@@ -70,6 +72,7 @@ export class LineChartComponent implements AfterViewInit, OnChanges, OnDestroy {
 
   public constructor(private changeDetectorRef: ChangeDetectorRef) {
     Chart.register(
+      Decimation,
       Filler,
       LineController,
       LineElement,
@@ -194,6 +197,9 @@ export class LineChartComponent implements AfterViewInit, OnChanges, OnDestroy {
             },
             interaction: { intersect: false, mode: 'index' },
             plugins: <unknown>{
+              decimation: {
+                enabled: this.dataDecimation
+              },
               legend: {
                 align: 'start',
                 display: this.showLegend,